mktests.sh: Introduce the notion of sandboxes.
There are certain cases where we need to purposely generate
an error (such as EACCES, etc). We therefore introduce a sandbox/
directory which shall exist in every test case directory that needs
one. This need is expressed by the presence of a `need-sandbox'
dummy file in the respective test case directory.